fix incompatible u-root module dep
Since u-root has some semver tag >v2.x but does not have a corresponding
go.mod file (reflecting >v2.x), u-root is marked as 'incompatible' in
the go.mod file. This is no problem in the first place but as soon as
you want to upgrade to a commit newer than the latest semver tag things
get messed up.

This commit adds a replace directive to the go.mod file to replace the
version of u-root with a pseudo version. Usually this can be done without
a replace directive but as mentioned, u-root does not use go modules
in a correct way yet.

You can use 'go get github.com/u-root/u-root@<commit-hash>' which will
fail but gives you the corresponding pseudo semver. Thies pseudo version
is then used on the right side of the replace directive.

Signed-off-by: Jens Drenhaus <jens.drenhaus@9elements.com>