retry mount, instead of failing immediately
When booting from an USB-stick, the STBOOT and STDATA partition wont
be available immediately, and as a result stboot will abort and reset
the system. The problem is at the kernel has not reached the point,
where the USB-stick can be mounted.

To work-around this, add a retry mechanism around the mount function.

Signed-off-by: Björn Töpel <bjorn@mullvad.net>