Make {host,security}_configuration.json parsing stricter.
When configuring stboot via the JSON files, all fields should be
always be explicitly typed out. If not, this should be considered an
error. This improves the experience for the user in the following
ways:
 * A user will be aware of the available configuration options
 * It is harder for a user to do misconfiguration

The JSON tests were rewritten, but keeping the old tests. New tests to
check for missing fields were added.

Signed-off-by: Björn Töpel <bjorn@mullvad.net>