network: Debug printouts again (#85)
* Print list of bonded interfaces properly

I wish I knew how to map() or do list comprehensions.

* Remove spurious debug printout

The MAC address of a newly created bonding device is not very
interesting to a user. This is an artifact from when we were debugging
MAC changing.

The creation of the MAC itself isn't very interesting either, so let's
remove the whole line.